third_party/blink/web_tests/platform/linux-chrome/external/wpt/touch-events/touch-globaleventhandler-interface-expected.txt

This is a testharness.js-based test.
[PRECONDITION_FAILED] Touch events in GlobalEventHandlers
  'expose legacy touch event APIs'
[PRECONDITION_FAILED] Touch events are GlobalEventHandlers' own property
  'expose legacy touch event APIs'
[PASS] Touch events not in Element.prototype
Harness: the test ran to completion.
